Moved gui rendering stuff in a standalone class
